Hooked up UI to controller/orchestrator

+++ src/SFA.DAS.ProviderApprenticeshipsService.Web/Controllers/CommitmentController.cs
@@ -79,7 +79,7 @@ namespace SFA.DAS.ProviderApprenticeshipsService.Web.Controllers
         }
 
         [HttpGet]
-        [Route("{hashedCommitmentId}/Edit/{hashedApprenticeshipId}")]
+        [Route("{hashedCommitmentId}/Edit/{hashedApprenticeshipId}", Name = "EditApprenticeship")]
         public async Task<ActionResult> Edit(long providerId, string hashedCommitmentId, string hashedApprenticeshipId)
         {
             var model = await _commitmentOrchestrator.GetApprenticeship(providerId, hashedCommitmentId, hashedApprenticeshipId);
@@ -116,22 +116,36 @@ namespace SFA.DAS.ProviderApprenticeshipsService.Web.Controllers
 
         [HttpGet]
         [Route("{hashedCommitmentId}/{hashedApprenticeshipId}/Delete")]
-        public ActionResult DeleteConfirmation(long providerId, string hashedCommitmentId, string hashedApprenticeshipId)
+        [OutputCache(CacheProfile = "NoCache")]
+        public async Task<ActionResult> DeleteConfirmation(long providerId, string hashedCommitmentId, string hashedApprenticeshipId)
         {
-            var viewModel = _commitmentOrchestrator.GetDeleteConfirmationModel(providerId, hashedCommitmentId, hashedApprenticeshipId);
+            var viewModel = await _commitmentOrchestrator.GetDeleteConfirmationModel(providerId, hashedCommitmentId, hashedApprenticeshipId);
 
             return View(viewModel);
         }
 
         [HttpPost]
+        [ValidateAntiForgeryToken]
+        [Route("{hashedCommitmentId}/{hashedApprenticeshipId}/Delete")]
         public async Task<ActionResult> DeleteConfirmation(DeleteConfirmationViewModel viewModel)
         {
             if (!ModelState.IsValid)
             {
-                View(viewModel);
+                return View(viewModel);
             }
 
-            await _commitmentOrchestrator.DeleteApprenticeship(viewModel);
+            if (!viewModel.DeleteConfirmed.Value)
+            {
+                return RedirectToRoute("EditApprenticeship", new
+                {
+                    providerId = viewModel.ProviderId,
+                    hashedCommitmentId = viewModel.HashedCommitmentId,
+                    hashedApprenticeshipId = viewModel.HashedApprenticeshipId
+                });
+            }
+
+            var deletedApprenticeshipName = await _commitmentOrchestrator.DeleteApprenticeship(viewModel);
+            SetInfoMessage($"Apprentice record for {deletedApprenticeshipName} deleted");
 
             return RedirectToRoute("CohortDetails", new { providerId = viewModel.ProviderId, hashedCommitmentId = viewModel.HashedCommitmentId });
         }
